fix #4311: add quarantine help-page for end-users
Having a short explanation of what the quarantine interface provides,
in the end-user interface should help users who wonder why they get
mails from some system, and what quarantine means. Instead of adding
this to the regular documentation and linking it in the top-bar as we
do in the admin-view having a short description directly rendered on
the main panel should increase the chances of it getting read.

the code is inspired by the notesview (I found it when clicking
through a PBS system of mine looking for how to address this).

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>