html-encode more text-based object-grid rows
To avoid visual glitches that can happen if a text value is wrongly
interpreted as html. And while quite a few rows already had a rather
restricted format that avoided this in the first place it still
doesn't hurt to encode anyway, that also protects against the case
where a backend format becomes less restrictive in the future.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>