dkim selector view: fix viewing larger DKIM TXT records
It seems the 'grow' feature of the textarea only works properly on user
input, not when automatically filling it with data. So instead of trying
to do that, which only ever clipped the record if it was too large,
simply use a fixed size that is enough for 4096 bit sized keys and show
a scrollbar for larger ones.

With having a fixed height, and horizontal resizing a textarea also not
working properly in an edit window, prevent the whole window from
resizing.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>