login view: fix realm change handler
record here can also be 'null', so it's not enough to check for
undefined. While at it, change the statement to conform to our style
guide.

This also fixes the quarantine link popup, because an exception in
a change handler interrupts normal extjs operation, which leads
to an invalid extjs state here.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>