login view: hide realm selector on quarantine view
when we target the quarantine view, we don't want to show the realm
field, since the only valid users here are from LDAP and that is not a
realm in that sense.

We do this by moving the realmfield hide/disable up before the
autologin, but after the targetview check.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>