fix #3284: improve display & editing of `dnsbl_sites`
Improve the display of the dnsbl_sites property by splitting on our
usual characters (',', ';', ' ') and put one entry per line.
For this to better work, put the DNSBL options at the bottom.

To improve the editing of this setting, introduce a DNSBLSitesGrid,
inspired from the tag color override grid from PVE. (There might be some
code to be shared with that, but the most interesting parts are different:
get/setValue, columns, etc.).

This allows to add/edit/remove single entries from the list much easier.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>