quarantine view: redirect unmatched route event to quarantine overview
The end-user quarantine overview profits from having safe-states in
case the navigation #hash is pointing to a unknown route, especially
as we only got a handful of routes there anyway.

So avoid leaving the UI in a potential undefined state but rather
navigate to the 'pmgSpamQuarantine' route, which is the overview of
all current spam, which we also default too if there is no route at
all.

Log the unmatched route in the developer console to make debugging
this easier though.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>