quarantine view: show toast with link to mobile view if viewport width <= 1100px
On smaller view port sizes show a toast for 7s with a button-link to
switching to the mobile view to cover all smaller devices, like
tablets or also really small laptops or other tiny screens.

Use a relative link target that just sets the mobile=1 get parameter
that the backend knows about, this should even work if we're behind
reverse proxies or the like, as long as GET parameters are forwarded.

The 1100px was chosen empirically by trying a few different sizes to
see when the quarantine UI really starts to become unusable.
I actually had 1200px as boundary first, but I lowered it for now,
tablets and the likes should still show it while not annoying those
that just have the browser window reduced to a smaller size by
default. We can still change this boundary, and while not really
problematic, growing it upwards (to also show it on bigger screens) is
here IMO safer than reducing it in the future.

The 7s were chosen to give enough time to react while still vanishing
relatively soon, make the toast actively closable for those that want
to be in control earlier.

If it annoys people we can add a "do not show again checkbox" that
saves the state in the local browser storage.

Partially-Fixes: https://bugzilla.proxmox.com/show_bug.cgi?id=3434
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>